shared/pyproject.toml
[project]
name = "microservices-shared"
version = "0.1.0"
description = "Shared logging utilities for the microservices logging demo."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-exporter-otlp-proto-http==1.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]
